external SMA trigger
This commit, I fixed the logic that latches on the SMA external trigger. To use this trigger method, in software you should set the configuration "wait for sys" using the 0x7 trigger settings lo command. This was tested in conjunction with a self trigger on the acdc. How this trigger method works when varying the options in the 0x7 "hi" block is untested (for example, varying how long the ACDC should hold the event for)