No-FIFO
This is a commit after Eric and Evan's second session in EDG. Eric essentially removed the FIFO that appears after the decoding and before writing to RAM and writes data straight to RAM from the decoder, then sends over uart.

A change was also made to the transcievers state machine to properly handle the WRITE_ADDRESS and WRITE_ENABLE variables which were not being toggled.

We have verified that bytes from ACC signal tap are being properly transferred to the software end. At the moment the software is not parsing these bytes in the correct order - so something is inconsistent.